requirements.txt
openai
requests
beautifulsoup4
python-dotenv
tenacity
loguru
mypy
fastapi
uvicorn
logtail-python
